Refactor:LogoutModal 로직 리팩토링
