+++ app/common/components/ui/topBar/PageTopBar/PageTopBar.tsx
@@ -2,7 +2,7 @@
 
 import MenuIcon from "@/common/assets/icons/menu/MenuIcon";
 import LogoutModal from "@/feature/auth/components/Logout/organisms/LogoutModal";
-import UseLogoutModal from "@/feature/auth/hooks/useLogoutModal";
+import UseLogoutModal from "@/feature/auth/hooks/Logout/useLogoutModal";
 import { twMerge } from "tailwind-merge";
 
 interface PageTopBarProps {
