+++ app/feature/auth/hooks/Form/useForm.tsx
@@ -1,5 +1,5 @@
 import { FormEvent, useState } from "react";
-import { LoginFormState, RegisterFormState } from "../type";
+import { LoginFormState, RegisterFormState } from "../../type";
 import { validateForm } from "@/common/utils/validate/validateForm";
 
 const FIELD = ["email", "password", "nickname"];
